requirements.txt

opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pandas==2.1.4
numpy==1.24.4
Pillow==10.1.0

Manual Installation:
- Run these commands one by one:

    pip install opencv-python
    pip install opencv-contrib-python  
    pip install pandas
    pip install numpy
    pip install Pillow